start labelling figures and atomizing scripts
